Added support for anchors on sidebar
I have added the support to add anchors on sidebar throught the sidebars.json file. I have added an optional key “anchors” for each category. The anchors key must be an array of string where each string is an anchor. The next step is to make this compatible with i18n.