Add doctype to generated html files.
Actually all generated html files doesn't have a doctype.
In order to fix this, a doctype has been added when renderToStaticMarkup
function is invoked.
Accordingly to theses changes, a test has been added in order
to test that all html files contains a doctype.
BxooeforeAll function has been updated to in order to ensure
that we targeted all html files, even ones which are in subdirectory
like /docs/en/*.html.