[RFC] Allow configure what version is shown by default
Seeing if this will work - it won't work if you want to show `master` or `next` by default.
Only named version that would exist in `versioned_docs`.

If this does work and we want to go with it - `latestVersion` in many places seems like a
misnomer. `defaultVersion` would be better or something