Make it clear where the `docs` folder is
It was not clear, to the beginner user—who this tutorial is for—where the `docs` folder was . The only reason I know this is because I'm a beginner user and I tried for too many minutes to find the `docs` folder inside the `website` folder. I had this assumption because the previous example is offered under the assumption that you're in the `website` folder.

Feel free to change the wording, I just want to make it clear where you should be looking, if you're new.