docs: add an HTTPS example for publish-gh-pages
I moved step 1 to step 2 so the terminal steps would be together. I added Git Bash, because `publish-gh-pages` worked for me in Git Bash but didn't work at the Windows command prompt. 

I added HTTPS to step 4 because, on my system, HTTPS worked for me but SSH didn't work. I changed <GIT_USER> to USERNAME and deleted the $ prompt, to be consistent with other examples in the tutorial.