fix (v1): Fix broken relative links related to `--skip-next-release` build option
The problem is described in issue #1805 and it seems to have existed for a long time.
Given versioning is enabled and the repository contains versioned docs, if the build
option `--skip-next-release` is used when building the static site, the versioned
docs contains a lot of broken links. The generator is not able to locate the targetted
file for the relative link and so it leaves the html links as ".md" instead of ".html".

The issue appears to be isolated to `website/versioned-docs/` only and only if the above
mentioned build option is used.

This change fixes this issue and allows for generating site with appropriate relative
links.