docs(v2): fix changelog require.resolve mention
* Update change log

The 2.0.0-alpha56 Change log does not make mention to the breaking change with plugin imports. The [GitHub release](https://github.com/facebook/docusaurus/releases/tag/v2.0.0-alpha.56) does. This has meant that dependabot has missed this change in it's pull requests (which seemed to use the changelog in some cases)

* Fix typo

Was in the source I copied from GH release