fix: resolve markdown links wtih encoded spaces
In order to create markdown links to URIs containing spaces, an
encoded space (%20) must be used. These types of links were not
properly resolved to doc files. This fix allows them to be
resolved by calling `decodeURI` on the links found in markdown files to
unescape characters such as spaces.