docs: Minor `gh-pages` deploy config improvements
Improve some comments and clarifies the file path beyond file name alone for each config file.

Additionally removes the workflow path triggers as in practice these being updated shouldn't be triggering a re-run of the workflow again (assuming deterministic build from same input results in same output).

If there is a need for such a manual trigger of the workflow is probably a better approach. Performing a build because workflow comments were modified only would be pointless for example.