Add documentation to add custom admonitions
Hello,
I found myself needing to add a custom admonition, and thought it would be worth adding documentation on how to do so. 

I've considered adding the admonition itself to this repository to be able to preview it in the docs, but I didn't want to propose a config change unless you agree beforehand.