Added more directions for Blog-Only Mode
The original docs did not explicitly explain that it is possible to keep a homepage and alter the name of the route for the blog. There was also no recommendations to adjust the default links in the navbar/footer. I also added in some behavior that is expected if they do not delete their home index.js and duplicate routes are found.