Update prefetch.ts
In your code snippet, the empty arrow function is being used as a catch handler for the supportedPrefetchStrategy promise. Instead of having an empty catch handler, it's better to provide a meaningful error message or handle the error in some way.
By logging the error to the console, you can get more information about what went wrong when the promise is rejected. This can be helpful when debugging your code.