fix: Give priority to props on Tabs
By switching the priority of inner properties (`...tabs`) with input properties (`...props`), we allow the users to control the `selectedValue` (i.e. switch tabs), as well as to pass a hook `selectValue` to act on a tab switch.

This is a minor change that should improve UX while not changing anything else.