SearchPage respects contextualSearch config
With this update the SearchPage component now checks the contextualSearch value of the config file. Disjunctive Facet Refinements are no longer added to the query if this setting is false.

Additionally if a user has now explicitly defined a title or description in their algolia index config, those items will be used in place of hierarchy items. otherwise hierarchy functions as normal.

Tested working for contextualSearch true & false.