Update search.mdx - insights:true
We (Algolia) recommend that DocSearch users utilize insights by default now. This will give DocSearch customers better insights out of the box into what users are doing with their searches, without having to turn it on later and then wait some period of time to get good data. It's OK if the default is still false, if you prefer, we recommend it to be true. If desired, you could change the code to have it true by default and then mention here it can be turned off.