fix(website): ensure feedback widget renders with correct theme
If a user lands on the feature request community page with the theme set to dark the canny
widget is rendered twice, first light and then dark. The first call with the light theme takes
effect as the widget has not yet loaded and the widget shows in light theme. Adding a small
delay prevents the first render with the wrong theme.